refactor : game 패키지에서 number 분리
