+++ src/main/java/baseball/game/BaseballGame.java
@@ -7,9 +7,9 @@ import java.util.List;
 import static baseball.constant.GameConstant.*;
 import static baseball.constant.GameMessage.*;
 import static baseball.game.CompareNumbers.*;
-import static baseball.game.ComputerNumber.createComputerNumber;
+import static baseball.number.ComputerNumber.createComputerNumber;
 import static baseball.exception.RestartNumberValidator.checkValidationRestartNumber;
-import static baseball.game.UserNumber.checkInputUserNumber;
+import static baseball.number.UserNumber.checkInputUserNumber;
 
 public class BaseballGame {
