+++ src/test/java/baseball/game/ComputerNumberTest.java
@@ -1,10 +1,10 @@
 package baseball.game;
 
+import baseball.number.ComputerNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class ComputerNumberTest {
